+++ cmd/frontend/internal/app/pkg/updatecheck/client.go
@@ -115,7 +115,7 @@ func updateURL(ctx context.Context) string {
 }
 
 func authProviderTypes() []string {
-	ps := conf.AuthProviders()
+	ps := conf.Get().AuthProviders
 	types := make([]string, len(ps))
 	for i, p := range ps {
 		types[i] = conf.AuthProviderType(p)